Exit on error
Avoid failing silently - which allowed a bug to creep in.
Sometimes the gofmt command will error because of something in the code it's
analyzing, like a parsing error.

In this case, it's good to blow up because the analysis doesn't continue.
In this case we should add the gofmt output to stderr for more information.